component getQuotes from API created
