+++ src/App.js
@@ -5,24 +5,12 @@ import Home from "./pages/Home";
 import Contact from "./pages/Contact";
 import Ask from "./pages/Ask.js";
 import History from "./pages/History";
-import axios from 'axios';
+import GetQuote from "./assets/components/getQuote/getQuote";
+// import axios from 'axios';
+// import { useState } from "react";
 
 
 function App() {
-
-// API call for random quotes
-  const getQuote = () => {
-    axios.get('https://api.quotable.io/random')
-    .then((response) => {
-    console.log(response.data.content);
-    console.log(response.data.author);
-})
-    .catch((error) => {
-    console.log(error);
-})
-  }
-
-
   
   return (
     <Router>
@@ -34,7 +22,7 @@ function App() {
         <Route path="/ask" element={<Ask />} />
         <Route path="/history" element={<History />} />
       </Routes>
-      <button onClick={getQuote}>Get Quote</button>
+      <GetQuote />
     </Router>
   );
 }
